working pretty print function
