+++ .vscode/arduino.json
@@ -2,5 +2,5 @@
     "board": "esp32:esp32:esp32",
     "configuration": "PSRAM=disabled,PartitionScheme=default,CPUFreq=240,FlashMode=qio,FlashFreq=80,FlashSize=4M,UploadSpeed=921600,DebugLevel=none",
     "port": "COM4",
-    "sketch": "examples\\Object_array\\Object_array.ino"
+    "sketch": "examples\\Nested\\Nested.ino"
 }
