+++ LICENSE
@@ -1,6 +1,10 @@
 The MIT License (MIT)
 
-Copyright (c) 2013 Tim
+Copyright (c) 2013 Tim - original ArduinoJsonWriter
+
+Copyright (c) 2020 Robert van den Breemen 
+    - adding pretty printing
+    - fix to enable reuse of object 
 
 Permission is hereby granted, free of charge, to any person obtaining a copy of
 this software and associated documentation files (the "Software"), to deal in
