+++ README.md
@@ -2,6 +2,8 @@ ArduinoJsonWriter
 =================
 
 November 2020: RvdB: This is fork from the orginal... adding some stuff I need. 
+- fixing the class so the object can be used in loops (begin/end document)
+- adding pretty print 
 
 Json Writer is a light weight library that make it easier to generate json document on Arduino.
