+++ library.json
@@ -1,11 +1,11 @@
 {
-    "name": "Json Writer",
-    "keywords": "json, writer, generator",
-    "description": "Json Writer is a light weight library that make it easier to generate json document on Arduino without memory limitation",
+    "name": "Imrpoved Json Writer",
+    "keywords": "json, writer, generator, pretty",
+    "description": "Json Writer is a light weight library that make it easier to generate (pretty) json document on Arduino without memory limitation",
     "repository":
     {
         "type": "git",
-        "url": "https://github.com/maxpowel/ArduinoJsonWriter.git"
+        "url": "https://github.com/rvdbreemen/ArduinoJsonWriter.git"
     },
     "frameworks": "arduino",
     "platforms": "atmelavr,atmelsam"
